requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
streamlit==1.38.0
gradio==4.44.0
altair==5.3.0
plotly==5.24.1
langchain==0.2.15
langchain-community==0.2.12
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
rank-bm25==0.2.2
tqdm==4.66.5
aiohttp==3.10.5
# FIX: downgrade protobuf to <5 for Google GenAI compatibility
protobuf>=4.25.3,<5.0.0
google-generativeai==0.3.2
